pyproject.toml
[project]
name = "cow-modbus"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael Zaramensky",email = "m.zaramensky@tochka.ai"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymodbus (>=3.9.2,<4.0.0)",
    "pyserial (>=3.5,<4.0)",
    "bitarray (>=3.5.1,<4.0.0)",
    "loguru   (>=0.7.2)"
]

[tool.poetry]
packages = [{include = "cow_modbus", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
